python-reader/requirements.txt
pyserial==3.4
mysql-connector==2.1.6